+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
- - add openapi.yml
+ - Add EPE survey formtypes 0003 and 0004 (147)
+ - Add openapi.yml
 
 ### 2.17.0 2020-06-09
  - Add construction survey formtype 0002 (228)